refactor: split ring validators into scene_checks_ring
scene_checks.py had grown to 1175 lines, past the 1000-line module cap.
Move the ring_system block's checkers and key inventories into the new
sibling spindoctor.sim.scene_checks_ring, exactly the way the bodies
entry checkers live in scene_checks_body: the shared scalar primitives
stay in scene_checks, the drivers in scene.py import the block checker
from its new home, and the editor coverage test reads the ring key
inventories from the new module.  scene_checks.py is now 752 lines and
scene_checks_ring.py 453.
